Questions and their relevant choices are fetched from the DB
Fetch API is used in JS to get the questions and their appropriate options depending on the clicked pagination control, without page reload. The class of the triggered pagination control is turned active each time it is clicked, and the class is removed when another is clicked, allowing the user to see the current page.